refactor(dashboard): decompose keeper-detail-panels.ts into flat sub-modules
keeper-detail-panels.ts was 1239 lines, well above the 300-line threshold.
Split into 8 focused flat modules following the fsm-hub-* precedent:

- keeper-detail-ctx-utils.ts    (~83 lines)  - pure utilities
- keeper-detail-kpi.ts          (~254 lines) - KPI primitives
- keeper-detail-outcomes.ts     (~133 lines) - OutcomesLedger
- keeper-detail-charts.ts       (~268 lines) - ContextChart, TokenTrendChart, MetricsCharts
- keeper-detail-telemetry.ts    (~275 lines) - Prompt/Inference telemetry panels
- keeper-detail-ctx-composition.ts (~176 lines) - CtxCompositionPanel
- keeper-detail-debug.ts        (~79 lines)  - RawDataDebug
- keeper-detail-lists.ts        (~52 lines)  - Equipment/Relationship/Traits lists

Original keeper-detail-panels.ts reduced to ~40-line barrel re-export
for backward compatibility. Import paths in agent-profile.ts and the
single test file updated accordingly.

Tests: 12 passed (keeper-detail-panels.test.ts)
Type-check: no new errors introduced
